fix gcc warnings

+++ tests/imgqsort/detectblur/detectblur.cpp
@@ -41,7 +41,7 @@ Mat src, src_gray;
 Mat dst, detected_edges;
 
 int  edgeThresh            = 1;
-int  lowThreshold          = 0.4;   //given in research paper
+double    lowThreshold     = 0.4;   //given in research paper
 int const max_lowThreshold = 100;
 int  ratio_value           = 3;
 int  kernel_size           = 3;
